third_party/blink/web_tests/fast/filesystem/op-move-expected.txt

* Running: MoveFileSimple
PASS Succeeded: "/a/b".moveTo("/a", "c")
PASS Verified entry: ENTRY {path:/a name:a type:DIRECTORY}
PASS Verified entry does NOT exist: /a/b
PASS Verified entry: ENTRY {path:/a/c name:c type:FILE}
* Running: MoveDirectorySimple
PASS Succeeded: "/a/b".moveTo("/a", "c")
PASS Verified entry: ENTRY {path:/a name:a type:DIRECTORY}
PASS Verified entry does NOT exist: /a/b
PASS Verified entry: ENTRY {path:/a/c name:c type:DIRECTORY}
* Running: MoveFileToDifferentDirectory
PASS Succeeded: "/a/b".moveTo("/c", "d")
PASS Verified entry: ENTRY {path:/a name:a type:DIRECTORY}
PASS Verified entry does NOT exist: /a/b
PASS Verified entry: ENTRY {path:/c/d name:d type:FILE}
* Running: MoveFileWithEmptyName
PASS Succeeded: "/a/b".moveTo("/c", "null")
PASS Verified entry: ENTRY {path:/a name:a type:DIRECTORY}
PASS Verified entry does NOT exist: /a/b
PASS Verified entry: ENTRY {path:/c/b name:b type:FILE}
* Running: MoveFileWithEmptyNameToSameDirectory
PASS this.expectedError + "" is "InvalidModificationError"
PASS Verified entry: ENTRY {path:/a name:a type:DIRECTORY}
PASS Verified entry: ENTRY {path:/a/b name:b type:FILE}
* Running: MoveFileWithSameName
PASS this.expectedError + "" is "InvalidModificationError"
PASS Verified entry: ENTRY {path:/a name:a type:DIRECTORY}
PASS Verified entry: ENTRY {path:/a/b name:b type:FILE}
* Running: MoveForNonExistentEntry
PASS Succeeded: "/a/b".remove()
PASS this.expectedError + "" is "NotFoundError"
PASS Verified entry: ENTRY {path:/a name:a type:DIRECTORY}
PASS Verified entry: ENTRY {path:/c name:c type:DIRECTORY}
* Running: MoveEntryToNonExistentDirectory
PASS Succeeded: "/c".remove()
PASS this.expectedError + "" is "NotFoundError"
PASS Verified entry: ENTRY {path:/a name:a type:DIRECTORY}
PASS Verified entry: ENTRY {path:/a/b name:b type:FILE}
* Running: MoveEntryToItsChild
PASS this.expectedError + "" is "InvalidModificationError"
PASS this.expectedError + "" is "InvalidModificationError"
PASS Verified entry: ENTRY {path:/a name:a type:DIRECTORY}
PASS Verified entry: ENTRY {path:/a/b name:b type:DIRECTORY}
PASS Verified entry: ENTRY {path:/a/b/c name:c type:DIRECTORY}
* Running: MoveRecursive
PASS Succeeded: "/a".moveTo("/b", "a")
PASS Verified entry does NOT exist: /a
PASS Verified entry: ENTRY {path:/b/a name:a type:DIRECTORY}
PASS Verified entry: ENTRY {path:/b/a/b name:b type:DIRECTORY}
PASS Verified entry: ENTRY {path:/b/a/b/c name:c type:FILE}
PASS Verified entry: ENTRY {path:/b/a/b/d name:d type:FILE}
* Running: OverwritingMoveFileToFile
PASS Succeeded: "/a".moveTo("/", "b")
PASS Verified entry: ENTRY {path:/b name:b type:FILE}
* Running: OverwritingMoveDirectoryToEmptyDirectory
PASS Succeeded: "/a".moveTo("/", "c")
PASS Verified entry: ENTRY {path:/c name:c type:DIRECTORY}
PASS Verified entry: ENTRY {path:/c/b name:b type:FILE}
PASS Verified entry does NOT exist: /a
* Running: OverwritingMoveFileToDirectory
PASS this.expectedError + "" is "InvalidModificationError"
PASS Verified entry: ENTRY {path:/a name:a type:FILE}
PASS Verified entry: ENTRY {path:/b name:b type:DIRECTORY}
* Running: OverwritingMoveDirectoryToFile
PASS this.expectedError + "" is "InvalidModificationError"
PASS Verified entry: ENTRY {path:/a name:a type:DIRECTORY}
PASS Verified entry: ENTRY {path:/b name:b type:FILE}
* Running: OverwritingMoveFileToNonemptyDirectory
PASS this.expectedError + "" is "InvalidModificationError"
PASS Verified entry: ENTRY {path:/a name:a type:FILE}
PASS Verified entry: ENTRY {path:/b name:b type:DIRECTORY}
PASS Verified entry: ENTRY {path:/b/c name:c type:FILE}
* Running: OverwritingMoveDirectoryToNonemptyDirectory
PASS this.expectedError + "" is "InvalidModificationError"
PASS Verified entry: ENTRY {path:/a name:a type:DIRECTORY}
PASS Verified entry: ENTRY {path:/a/b name:b type:FILE}
PASS Verified entry: ENTRY {path:/c name:c type:DIRECTORY}
PASS Verified entry: ENTRY {path:/c/d name:d type:FILE}
Finished running tests.
PASS successfullyParsed is true

TEST COMPLETE
